# packages/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EquiDuty API",
    "description": "# EquiDuty API Documentation\n\nComprehensive REST API for stable management, horse tracking, scheduling, and billing.\n\n## Authentication\nAll endpoints require Firebase Auth JWT token in Authorization header:\n```\nAuthorization: Bearer <firebase-jwt-token>\n```\n\n## Authorization\n- **System Roles**: system_admin, stable_owner, stable_user\n- **Organization Roles**: 18 roles including administrator, groom, rider, veterinarian, etc.\n- **Permissions**: 60+ granular actions via Permission Engine V2\n- **Field-Level Access**: Horse data projection based on access levels (public → owner)\n\n## Response Format\n- **Success**: Serialized Firestore data (timestamps converted to ISO strings)\n- **Error**: `{ error: string, message: string, details?: unknown }`\n\n## Rate Limiting\n- Standard: 100 requests/minute per user\n- Burst: 20 requests/second per user\n\n## Environments\n- **Development**: https://equiduty-dev.ew.r.appspot.com/api/v1\n- **Staging**: https://equiduty-staging.ew.r.appspot.com/api/v1\n- **Production**: https://api.equiduty.com/api/v1",
    "version": "1.0.0",
    "contact": {
      "name": "EquiDuty Support",
      "email": "support@equiduty.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase Auth JWT token. Obtain via Firebase Authentication SDK."
      }
    },
...
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Complete user registration by creating Firestore profile, personal organization with implicit stable, and migrating pending invites. Firebase Auth user must be created on frontend first.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "firstName",
                  "lastName"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "firstName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "organizationType": {
                    "type": "string",
                    "enum": [
                      "personal",
                      "business"
                    ]
                  }
                }
              }
            }
          }
        },
        "x-permissions": {
          "authenticated": true,
          "systemRole": null,
          "organizationPermission": null,
          "stablePermission": null
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "description": "User successfully registered",
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "systemRole": {
                          "type": "string",
                          "enum": [
                            "stable_user"
                          ]
                        },
                        "createdAt": {
                          "type": "object"
                        },
                        "updatedAt": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or body",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request parameters or body",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Missing or invalid JWT token",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already registered",
            "content": {
              "application/json": {
                "schema": {
                  "description": "User already registered",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "Get authenticated user profile from Firestore",
        "x-permissions": {
          "authenticated": true,
          "systemRole": null,
          "organizationPermission": null,
          "stablePermission": null
        },
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "User profile retrieved successfully",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "systemRole": {
                      "type": "string",
                      "enum": [
                        "system_admin",
                        "stable_owner",
                        "stable_user"
                      ]
                    },
                    "createdAt": {
                      "type": "object"
                    },
                    "updatedAt": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Missing or invalid JWT token",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Resource not found",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/settings": {
      "patch": {
        "tags": [
          "Authentication"
        ],
        "description": "Update user preferences and notification settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emailNotifications": {
                    "type": "boolean"
                  },
                  "pushNotifications": {
                    "type": "boolean"
                  },
                  "darkMode": {
                    "type": "boolean"
                  },
                  "timezone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "x-permissions": {
          "authenticated": true,
          "systemRole": null,
          "organizationPermission": null,
          "stablePermission": null
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Settings updated successfully",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "settings": {
                      "type": "object",
                      "properties": {
                        "emailNotifications": {
                          "type": "boolean"
                        },
                        "pushNotifications": {
                          "type": "boolean"
                        },
                        "darkMode": {
                          "type": "boolean"
                        },
                        "timezone": {
                          "type": "string"
                        }
                      }
                    },
                    "updatedAt": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters or body",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Invalid request parameters or body",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Missing or invalid JWT token",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Resource not found",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://equiduty-dev.ew.r.appspot.com/api/v1",
      "description": "Development Environment"
    },
    {
...
    },
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}